IF Refactors and Rebuild
