+++ docs/iframe.html
@@ -130,4 +130,4 @@
             
         
             
-              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.3fe2a8f73757269eafea.bundle.js"></script><script src="vendors~main.3fe2a8f73757269eafea.bundle.js"></script><script src="main.3fe2a8f73757269eafea.bundle.js"></script></body></html>
+              window['FRAMEWORK_OPTIONS'] = {};</script><script src="runtime~main.d6de1ac81c9d957108be.bundle.js"></script><script src="vendors~main.d6de1ac81c9d957108be.bundle.js"></script><script src="main.d6de1ac81c9d957108be.bundle.js"></script></body></html>
