--- src/interfaces/contact-interface.ts
@@ -1,6 +1,0 @@
-export interface ContactIF{
-  email: string
-  confirmEmail?: string
-  phone: string
-  extension?: string
-}
