+++ src/interfaces/index.ts
@@ -1,8 +1,7 @@
 export * from './alteration-interface'
 export * from './business-interface'
-export * from './businessContact-interface'
+export * from './contactPoint-interface'
 export * from './comment-interface'
-export * from './contact-interface'
 export * from './filingData-interface'
 export * from './form-interface'
 export * from './nameRequest-interface'
